requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
ca-certificates=2021.5.30=ha878542_0
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
lazy-object-proxy=1.4.3=py37h7b6447c_0
ld_impl_linux-64=2.35.1=h7274673_9
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.3.0=hdf63c60_0
libstdcxx-ng=9.3.0=hd4cf53a_17
mkl-service=2.3.0=py37he904b0f_0
more-itertools=8.5.0=py_0
numpy-base=1.19.1=py37hfa32c7d_0
python-dotenv=0.19.0=pyhd8ed1ab_0
typed-ast=1.4.1=py37h7b6447c_0
